+++ src/daemon/mod.rs
@@ -10,18 +10,15 @@ use crate::blocks::Tipset;
 use crate::chain::HeadChange;
 use crate::chain_sync::ChainFollower;
 use crate::chain_sync::network_context::SyncNetworkContext;
-use crate::cli_shared::{car_db_path, snapshot};
+use crate::cli_shared::snapshot;
 use crate::cli_shared::{
     chain_path,
     cli::{CliOpts, Config},
 };
 use crate::daemon::context::{AppContext, DbType};
-use crate::daemon::db_util::{
-    import_chain_as_forest_car, load_all_forest_cars, populate_eth_mappings,
-};
-use crate::db::SettingsStore;
-use crate::db::car::ManyCar;
-use crate::db::{MarkAndSweep, MemoryDB, SettingsExt, ttl::EthMappingCollector};
+use crate::daemon::db_util::import_chain_as_forest_car;
+use crate::db::gc::{MarkAndSweep, SnapshotGarbageCollector};
+use crate::db::ttl::EthMappingCollector;
 use crate::libp2p::{Libp2pService, PeerManager};
 use crate::message_pool::{MessagePool, MpoolConfig, MpoolRpcProvider};
 use crate::networks::{self, ChainConfig};
@@ -30,7 +27,6 @@ use crate::rpc::eth::filter::EthEventHandler;
 use crate::rpc::start_rpc;
 use crate::shim::clock::ChainEpoch;
 use crate::shim::version::NetworkVersion;
-use crate::state_manager::StateManager;
 use crate::utils;
 use crate::utils::{
     monitoring::MemStatsTracker, proofs_api::ensure_proof_params_downloaded,
@@ -39,11 +35,12 @@ use crate::utils::{
 use anyhow::{Context as _, bail};
 use dialoguer::theme::ColorfulTheme;
 use futures::{Future, FutureExt, select};
-use fvm_ipld_blockstore::Blockstore;
 use once_cell::sync::Lazy;
 use raw_sync_2::events::{Event, EventInit as _, EventState};
 use shared_memory::ShmemConf;
 use std::path::Path;
+use std::sync::OnceLock;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
 use std::{cmp, sync::Arc};
 use tempfile::{Builder, TempPath};
@@ -66,6 +63,8 @@ static IPC_PATH: Lazy<TempPath> = Lazy::new(|| {
         .into_temp_path()
 });
 
+pub static GLOBAL_SNAPSHOT_GC: OnceLock<Arc<SnapshotGarbageCollector<DbType>>> = OnceLock::new();
+
 // The parent process and the daemonized child communicate through an Event in
 // shared memory. The identity of the shared memory object is written to a
 // temporary file. The parent process is responsible for cleaning up the file
@@ -78,13 +77,19 @@ pub fn ipc_shmem_conf() -> ShmemConf {
 }
 
 fn unblock_parent_process() -> anyhow::Result<()> {
-    let shmem = ipc_shmem_conf().open()?;
-    let (event, _) =
-        unsafe { Event::from_existing(shmem.as_ptr()).map_err(|err| anyhow::anyhow!("{err}")) }?;
+    static UNBLOCKED: AtomicBool = AtomicBool::new(false);
+    if !UNBLOCKED.load(Ordering::Relaxed) {
+        let shmem = ipc_shmem_conf().open()?;
+        let (event, _) = unsafe {
+            Event::from_existing(shmem.as_ptr()).map_err(|err| anyhow::anyhow!("{err}"))
+        }?;
 
-    event
-        .set(EventState::Signaled)
-        .map_err(|err| anyhow::anyhow!("{err}"))
+        event
+            .set(EventState::Signaled)
+            .map_err(|err| anyhow::anyhow!("{err}"))?;
+        UNBLOCKED.store(true, Ordering::Relaxed);
+    }
+    Ok(())
 }
 
 /// Increase the file descriptor limit to a reasonable number.
@@ -132,9 +137,6 @@ pub async fn start_interruptable(opts: CliOpts, config: Config) -> anyhow::Resul
     crate::utils::io::terminal_cleanup();
     result
 }
-
-// Garbage collection interval, currently set at 10 hours.
-const GC_INTERVAL: Duration = Duration::from_secs(60 * 60 * 10);
 
 /// This function initialize Forest with below steps
 /// - increase file descriptor limit (for parity-db)
@@ -274,12 +276,16 @@ async fn maybe_start_metrics_service(
     Ok(())
 }
 
-fn maybe_start_gc_service(
+#[allow(dead_code)]
+fn maybe_start_mark_and_sweep_gc_service(
     services: &mut JoinSet<anyhow::Result<()>>,
     opts: &CliOpts,
     config: &Config,
     ctx: &AppContext,
 ) {
+    // Garbage collection interval, currently set at 10 hours.
+    const GC_INTERVAL: Duration = Duration::from_secs(60 * 60 * 10);
+
     if !opts.no_gc {
         let mut db_garbage_collector = {
             let chain_store = ctx.state_manager.chain_store().clone();
@@ -392,7 +398,6 @@ async fn maybe_start_health_check_service(
                 .timestamp,
             sync_status: chain_follower.sync_status.clone(),
             peer_manager: p2p_service.peer_manager().clone(),
-            settings_store: ctx.db.writer().clone(),
         };
         let healthcheck_address = forest_state.config.client.healthcheck_address;
         info!("Healthcheck endpoint will listen at {healthcheck_address}");
@@ -467,12 +472,12 @@ fn maybe_start_rpc_service(
     Ok(())
 }
 
-fn maybe_start_f3_service(
-    services: &mut JoinSet<anyhow::Result<()>>,
-    opts: &CliOpts,
-    config: &Config,
-    ctx: &AppContext,
-) {
+fn maybe_start_f3_service(opts: &CliOpts, config: &Config, ctx: &AppContext) {
+    // already running
+    if crate::rpc::f3::F3_LEASE_MANAGER.get().is_some() {
+        return;
+    }
+
     if !config.client.enable_rpc {
         if crate::f3::is_sidecar_ffi_enabled(ctx.state_manager.chain_config()) {
             tracing::warn!("F3 sidecar is enabled but not run because RPC is disabled. ")
@@ -485,7 +490,7 @@ fn maybe_start_f3_service(
         let state_manager = &ctx.state_manager;
         let p2p_peer_id = ctx.p2p_peer_id;
         let admin_jwt = ctx.admin_jwt.clone();
-        services.spawn_blocking({
+        tokio::task::spawn_blocking({
             crate::rpc::f3::F3_LEASE_MANAGER
                 .set(crate::rpc::f3::F3LeaseManager::new(
                     state_manager.chain_config().network.clone(),
@@ -516,29 +521,7 @@ fn maybe_start_f3_service(
                     std::env::var("FOREST_F3_ROOT")
                         .unwrap_or(default_f3_root.display().to_string()),
                 );
-                Ok(())
             }
-        });
-    }
-}
-
-fn maybe_populate_eth_mappings_in_background(
-    services: &mut JoinSet<anyhow::Result<()>>,
-    opts: &CliOpts,
-    config: Config,
-    ctx: &AppContext,
-) {
-    if config.chain_indexer.enable_indexer
-        && !opts.stateless
-        && !ctx.state_manager.chain_config().is_devnet()
-    {
-        let state_manager = ctx.state_manager.clone();
-        let settings = ctx.db.writer().clone();
-        services.spawn(async move {
-            if let Err(err) = init_ethereum_mapping(state_manager, &settings, &config) {
-                tracing::warn!("Init Ethereum mapping failed: {}", err)
-            }
-            Ok(())
         });
     }
 }
@@ -595,13 +578,45 @@ fn maybe_start_indexer_service(
 pub(super) async fn start(
     start_time: chrono::DateTime<chrono::Utc>,
     opts: CliOpts,
-    mut config: Config,
+    config: Config,
     shutdown_send: mpsc::Sender<()>,
 ) -> anyhow::Result<()> {
     startup_init(&opts, &config)?;
+    let (snap_gc, snap_gc_reboot_rx) = SnapshotGarbageCollector::new(&config)?;
+    let snap_gc = Arc::new(snap_gc);
+    GLOBAL_SNAPSHOT_GC
+        .set(snap_gc.clone())
+        .ok()
+        .context("failed to set GLOBAL_SNAPSHOT_GC")?;
+    tokio::task::spawn({
+        let snap_gc = snap_gc.clone();
+        async move { snap_gc.event_toop().await }
+    });
+    loop {
+        tokio::select! {
+            _ = snap_gc_reboot_rx.recv_async() => {
+                snap_gc.cleanup_before_reboot().await;
+            }
+            result = start_services(start_time, &opts, config.clone(), shutdown_send.clone(), |ctx| {
+                snap_gc.set_db(ctx.db.clone());
+            }) => {
+                break result
+            }
+        }
+    }
+}
+
+pub(super) async fn start_services(
+    start_time: chrono::DateTime<chrono::Utc>,
+    opts: &CliOpts,
+    mut config: Config,
+    shutdown_send: mpsc::Sender<()>,
+    on_app_context_initialized: impl Fn(&AppContext),
+) -> anyhow::Result<()> {
     let mut services = JoinSet::new();
-    maybe_start_track_peak_rss_service(&mut services, &opts);
-    let ctx = AppContext::init(&opts, &config).await?;
+    maybe_start_track_peak_rss_service(&mut services, opts);
+    let ctx = AppContext::init(opts, &config).await?;
+    on_app_context_initialized(&ctx);
     info!(
         "Using network :: {}",
         get_actual_chain_name(&ctx.network_name)
@@ -611,10 +626,8 @@ pub(super) async fn start(
         return Ok(());
     }
     let p2p_service = create_p2p_service(&mut services, &mut config, &ctx).await?;
-
     let mpool = create_mpool(&mut services, &p2p_service, &ctx)?;
-
-    let chain_follower = create_chain_follower(&opts, &p2p_service, mpool.clone(), &ctx)?;
+    let chain_follower = create_chain_follower(opts, &p2p_service, mpool.clone(), &ctx)?;
 
     info!(
         "Starting network:: {}",
@@ -631,7 +644,7 @@ pub(super) async fn start(
         &ctx,
     )?;
 
-    maybe_import_snapshot(&opts, &mut config, &ctx).await?;
+    maybe_import_snapshot(opts, &mut config, &ctx).await?;
     if opts.halt_after_import {
         // Cancel all async services
         services.shutdown().await;
@@ -639,12 +652,11 @@ pub(super) async fn start(
     }
     ctx.state_manager.populate_cache();
     maybe_start_metrics_service(&mut services, &config, &ctx).await?;
-    maybe_start_gc_service(&mut services, &opts, &config, &ctx);
-    maybe_start_f3_service(&mut services, &opts, &config, &ctx);
+    // maybe_start_mark_and_sweep_gc_service(&mut services, opts, &config, &ctx);
+    maybe_start_f3_service(opts, &config, &ctx);
     maybe_start_health_check_service(&mut services, &config, &p2p_service, &chain_follower, &ctx)
         .await?;
-    maybe_populate_eth_mappings_in_background(&mut services, &opts, config.clone(), &ctx);
-    maybe_start_indexer_service(&mut services, &opts, &config, &ctx);
+    maybe_start_indexer_service(&mut services, opts, &config, &ctx);
     if !opts.stateless {
         ensure_proof_params_downloaded().await?;
     }
@@ -771,27 +783,4 @@ where
     T: Send + 'static,
 {
     tokio::task::spawn_blocking(f).then(|res| async { res.expect("spawned task panicked") })
-}
-
-fn init_ethereum_mapping<DB: Blockstore>(
-    state_manager: Arc<StateManager<DB>>,
-    settings: &impl SettingsStore,
-    config: &Config,
-) -> anyhow::Result<()> {
-    match settings.eth_mapping_up_to_date()? {
-        Some(false) | None => {
-            let car_db_path = car_db_path(config)?;
-            let db: Arc<ManyCar<MemoryDB>> = Arc::default();
-            load_all_forest_cars(&db, &car_db_path)?;
-            let ts = db.heaviest_tipset()?;
-
-            populate_eth_mappings(&state_manager, &ts)?;
-
-            settings.set_eth_mapping_up_to_date()
-        }
-        Some(true) => {
-            tracing::info!("Ethereum mapping up to date");
-            Ok(())
-        }
-    }
 }