+++ src/rpc/methods/chain.rs
@@ -195,6 +195,30 @@ impl RpcMethod<1> for ChainGetMessagesInTipset {
     }
 }
 
+pub enum ChainPruneSnapshot {}
+impl RpcMethod<1> for ChainPruneSnapshot {
+    const NAME: &'static str = "Forest.SnapshotGC";
+    const PARAM_NAMES: [&'static str; 1] = ["blocking"];
+    const API_PATHS: BitFlags<ApiPaths> = ApiPaths::all();
+    const PERMISSION: Permission = Permission::Admin;
+
+    type Params = (bool,);
+    type Ok = ();
+
+    async fn handle(
+        _ctx: Ctx<impl Blockstore + Send + Sync + 'static>,
+        (blocking,): Self::Params,
+    ) -> Result<Self::Ok, ServerError> {
+        if let Some(gc) = crate::daemon::GLOBAL_SNAPSHOT_GC.get() {
+            let progress_rx = gc.trigger()?;
+            while blocking && progress_rx.recv_async().await.is_ok() {}
+            Ok(())
+        } else {
+            Err(anyhow::anyhow!("snapshot gc is not enabled").into())
+        }
+    }
+}
+
 pub enum ChainExport {}
 impl RpcMethod<1> for ChainExport {
     const NAME: &'static str = "Filecoin.ChainExport";