+++ src/tool/subcommands/api_cmd/test_snapshot.rs
@@ -22,6 +22,7 @@ use crate::{
     shim::address::{CurrentNetwork, Network},
     state_manager::StateManager,
 };
+use cid::Cid;
 use openrpc_types::ParamStructure;
 use parking_lot::RwLock;
 use serde::{Deserialize, Serialize};
@@ -33,7 +34,8 @@ pub struct Payload(#[serde(with = "crate::lotus_json::base64_standard")] pub Vec
 
 #[derive(Default, PartialEq, Debug, Clone, Serialize, Deserialize)]
 pub struct Index {
-    pub eth_mappings: ahash::HashMap<String, Payload>,
+    pub eth_mappings: Option<ahash::HashMap<String, Payload>>,
+    pub indices: Option<ahash::HashMap<String, Payload>>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -51,9 +53,17 @@ pub struct RpcTestSnapshot {
 fn backfill_eth_mappings(db: &MemoryDB, index: Option<Index>) -> anyhow::Result<()> {
     if let Some(index) = index {
         if let Some(mut guard) = db.eth_mappings_db.try_write() {
-            for (k, v) in index.eth_mappings.into_iter() {
-                let hash = EthHash::from_str(&k)?;
-                guard.insert(hash, v.0);
+            if let Some(eth_mappings) = index.eth_mappings {
+                for (k, v) in eth_mappings.iter() {
+                    guard.insert(EthHash::from_str(k)?, v.0.clone());
+                }
+            }
+        }
+        if let Some(mut guard) = db.indices_db.try_write() {
+            if let Some(indices) = index.indices {
+                for (k, v) in indices.iter() {
+                    guard.insert(Cid::from_str(k)?, v.0.clone());
+                }
             }
         }
     }