+++ CHANGELOG.md
@@ -29,11 +29,19 @@
 
 ### Added
 
+-[#5598](https://github.com/ChainSafe/forest/pull/5598) Add `forest-cli chain prune snap` command for garbage collecting the database with a new snapshot garbage collector.
+
+-[#5629](https://github.com/ChainSafe/forest/pull/5629) Save default RPC token and consume it automatically.
+
 ### Changed
+
+-[#5616](https://github.com/ChainSafe/forest/pull/5616) Remove the initial background task for populating Ethereum mappings. Use `forest-tool index backfill` to perform this operation offline instead.
 
 ### Removed
 
 ### Fixed
+
+- [#5624](https://github.com/ChainSafe/forest/pull/5624) Fixed `Filecoin.EthTraceFilter` to correctly handle null tipsets and stay within the filter range.
 
 ## Forest v0.26.1 "Ijon Tichy"